ADD parsing of nimble files via nimble
Nimble files need to be loaded for this package.
That can be done via nimble itself, no need to reimplement that logic.
However, storing that is not a good idea,
as you want to keep your data as up to date as possible.
Therefore you should rather store path to nimble files and always
load the from disk